refactor: split output filename helper
