+++ packages/astro/src/core/util.ts
@@ -1,12 +1,10 @@
 import fs from 'node:fs';
-import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 import type { AstroSettings } from '../types/astro.js';
 import type { AstroConfig } from '../types/public/config.js';
-import type { RouteData } from '../types/public/internal.js';
 import { hasSpecialQueries } from '../vite-plugin-utils/index.js';
 import { SUPPORTED_MARKDOWN_FILE_EXTENSIONS } from './constants.js';
-import { removeQueryString, removeTrailingForwardSlash, slash } from './path.js';
+import { removeQueryString, slash } from './path.js';
 
 /** Check if a file is a markdown file based on its extension */
 export function isMarkdownFile(fileId: string, option?: { suffix?: string }): boolean {
@@ -19,33 +17,6 @@ export function isMarkdownFile(fileId: string, option?: { suffix?: string }): bo
 		if (id.endsWith(`${markdownFileExtension}${_suffix}`)) return true;
 	}
 	return false;
-}
-
-const STATUS_CODE_PAGES = new Set(['/404', '/500']);
-
-/**
- * Get the correct output filename for a route, based on your config.
- * Handles both "/foo" and "foo" `name` formats.
- * Handles `/404` and `/` correctly.
- */
-export function getOutputFilename(
-	buildFormat: NonNullable<AstroConfig['build']>['format'],
-	name: string,
-	routeData: RouteData,
-) {
-	if (routeData.type === 'endpoint') {
-		return name;
-	}
-	if (name === '/' || name === '') {
-		return path.posix.join(name, 'index.html');
-	}
-	if (buildFormat === 'file' || STATUS_CODE_PAGES.has(name)) {
-		return `${removeTrailingForwardSlash(name || 'index')}.html`;
-	}
-	if (buildFormat === 'preserve' && !routeData.isIndex) {
-		return `${removeTrailingForwardSlash(name || 'index')}.html`;
-	}
-	return path.posix.join(name, 'index.html');
 }
 
 /** is a specifier an npm package? */
